perf: do some simple performance optimization
