+++ packages/vgrammar-core/src/graph/attributes/helpers.ts
@@ -1,47 +1,8 @@
 import type { IPointLike } from '@visactor/vutils';
 // eslint-disable-next-line no-duplicate-imports
-import { isValidNumber, isString } from '@visactor/vutils';
+import { isValidNumber } from '@visactor/vutils';
 import type { MarkElementItem, MarkType } from '../../types';
-
-/**
- * 比较颜色是否相等
- * @param {*} c1
- * @param {*} c2
- * @returns
- */
-export function colorEqual(c1: any, c2: any) {
-  if (c1 === c2) {
-    return true;
-  }
-  if (!c1 || !c2) {
-    return false;
-  }
-  if (isString(c1)) {
-    return false;
-  }
-  if (['gradient', 'x0', 'x1', 'r0', 'r1', 'y0', 'y1'].some(key => c1[key] !== c2[key])) {
-    return false;
-  }
-  if (
-    c1.stops.length !== c2.stops.length ||
-    c1.stops.some((s: any, i: any) => s.offset !== c2.stops[i].offset || s.color !== c2.stops[i].color)
-  ) {
-    return false;
-  }
-  return true;
-}
-
-/**
- * 解析graphicItem x, y 相关默认值的处理
- * @param val
- * @returns
- */
-export function parseXY(val: number): number {
-  if (val === null) {
-    return 0;
-  }
-  return val;
-}
+import { GrammarMarkType } from '../enums';
 
 export function isValidPointsChannel(channels: string[], markType: MarkType): boolean {
   switch (markType) {
@@ -96,14 +57,14 @@ export function getLinePoints(
   }
   return items.map((item, index) => {
     const attrs = item.nextAttrs;
-    const x = parseXY(attrs.x ?? lastPoints?.[index]?.x);
-    const y = parseXY(attrs.y ?? lastPoints?.[index]?.y);
+    const x = attrs.x ?? lastPoints?.[index]?.x;
+    const y = attrs.y ?? lastPoints?.[index]?.y;
     const defined = attrs.defined ?? lastPoints?.[index]?.defined;
     const point: IPointLike = { x, y, context: item.key };
 
     if (isArea) {
-      const x1 = parseXY(attrs.x1 ?? lastPoints?.[index]?.x1);
-      const y1 = parseXY(attrs.y1 ?? lastPoints?.[index]?.y1);
+      const x1 = attrs.x1 ?? lastPoints?.[index]?.x1;
+      const y1 = attrs.y1 ?? lastPoints?.[index]?.y1;
       point.x1 = x1;
       point.y1 = y1;
     }
@@ -128,10 +89,10 @@ export function getLargeRectsPoints(
 
   items.forEach((item, index) => {
     const attrs = item.nextAttrs;
-    const x = parseXY(attrs.x ?? lastPoints?.[index * 4]);
-    const y = parseXY(attrs.y ?? lastPoints?.[index * 4 + 1]);
-    const width = parseXY(attrs.width ?? lastPoints?.[index * 4 + 2]);
-    const y1 = parseXY(attrs.y1 ?? lastPoints?.[index * 4 + 3]);
+    const x = attrs.x ?? lastPoints?.[index * 4];
+    const y = attrs.y ?? lastPoints?.[index * 4 + 1];
+    const width = attrs.width ?? lastPoints?.[index * 4 + 2];
+    const y1 = attrs.y1 ?? lastPoints?.[index * 4 + 3];
     arr[index * 4] = x;
     arr[index * 4 + 1] = y;
     arr[index * 4 + 2] = width;
@@ -153,11 +114,49 @@ export function getLargeSymbolsPoints(
 
   items.forEach((item, index) => {
     const attrs = item.nextAttrs;
-    const x = parseXY(attrs.x ?? lastPoints?.[index * 2]);
-    const y = parseXY(attrs.y ?? lastPoints?.[index * 2 + 1]);
+    const x = attrs.x ?? lastPoints?.[index * 2];
+    const y = attrs.y ?? lastPoints?.[index * 2 + 1];
     arr[index * 2] = x;
     arr[index * 2 + 1] = y;
   });
 
   return arr;
+}
+
+export function isPositionOrSizeChannel(type: string, channel: string) {
+  if (['x', 'y', 'dx', 'dy'].includes(channel)) {
+    return true;
+  }
+
+  switch (type) {
+    case GrammarMarkType.arc:
+      return ['innerRadius', 'outerRadius', 'startAngle', 'endAngle'].includes(channel);
+    case GrammarMarkType.group:
+    case GrammarMarkType.rect:
+    case GrammarMarkType.image:
+      return ['width', 'height', 'y1'].includes(channel);
+    case GrammarMarkType.path:
+    case GrammarMarkType.shape:
+      return ['path', 'customPath'].includes(channel);
+    case GrammarMarkType.line:
+      return channel === 'defined';
+    case GrammarMarkType.area:
+      return ['x1', 'y1', 'defined'].includes(channel);
+    case GrammarMarkType.rule:
+      return ['x1', 'y1'].includes(channel);
+    case GrammarMarkType.symbol:
+      return channel === 'size';
+    case GrammarMarkType.polygon:
+      return channel === 'points';
+    case GrammarMarkType.text:
+      return channel === 'text';
+  }
+
+  return false;
+}
+
+export function isPointsMarkType(markType: MarkType): boolean {
+  return (
+    [GrammarMarkType.line, GrammarMarkType.area, GrammarMarkType.largeRects, GrammarMarkType.largeSymbols] as MarkType[]
+  ).includes(markType);
 }
