+++ packages/vgrammar-core/src/interactions/player-filter.ts
@@ -1,5 +1,5 @@
 import type { DataFilterOptions, IComponent, IPlayer, IView, PlayerFilterValue } from '../types';
-import { DataFilterRank, GrammarMarkType } from '../graph';
+import { DataFilterRank, GrammarMarkType } from '../graph/enums';
 import { isString } from '@visactor/vutils';
 import { Filter } from './filter';
 import { PlayerEventEnum } from '@visactor/vrender-components';
