+++ packages/vgrammar-core/src/semantic-marks/text.ts
@@ -2,9 +2,10 @@ import { GrammarMarkType } from '../graph/enums';
 import type { IMark, MarkSpec } from '../types';
 import { Mark } from '../view/mark';
 import { createGraphicItem } from '../graph/util/graphic';
-import { transformsByType } from '../graph/attributes';
+import { transformsByType } from '../graph/attributes/transform';
 
 export class Text extends Mark {
+  static markType = GrammarMarkType.text;
   declare markType: GrammarMarkType.text;
   protected declare spec: MarkSpec;
